pyproject.toml
[project]
name = "studentconnect"
version = "1.0.0"
description = "A web app allowing the comparison of student timetables."
dependencies = [
	"flask",
	"PyJWT<2.10",
	"python-dotenv"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
